v1.2.7 - ALSession.ready() Helper
Added hooks to ALSession to allow external session checkers to signal
when they start and end.

Updated AlSessionDetector to use the startDetection() and endDetection()
methods in ALSession.

Added a new public method `ALSession.ready()` that is only resolved when
    1.  At least one session detection cycle has completed execution
    2.  No session detection cycles are currently in flight
    3.  The user is either unauthenticated OR the acting account has been resolved

Unlike `ALSession.resolved()`, which only resolves in authenticated
contexts, `ALSession.ready()` will resolve for unauthenticated users,
and indicates the stability of the session state.